+++ example/lib/presentation/basic_usage/view.dart
@@ -26,7 +26,7 @@ class _BasicUsageExampleState extends State<BasicUsageExample> {
             // Content behind the panel
             SlidingPanelBuilder(
               controller: controller,
-              snapConfig: SlidingPanelSnapConfig(extents: [0.75]),
+              snapConfig: SnapConfig(extents: [0.75]),
               handle: const SlidingPanelHandle(),
               builder: (context, handle) {
                 return SlidingPanelBody(