+++ example/lib/presentation/home/route/route.dart
@@ -4,6 +4,7 @@ import 'package:example/presentation/home/view.dart';
 import 'package:example/presentation/nested_scroll/route.dart';
 import 'package:example/presentation/adaptive_sizing/route.dart';
 import 'package:example/presentation/parallax_effect/route.dart';
+import 'package:example/presentation/transition_builder/route.dart';
 import 'package:flutter/widgets.dart';
 import 'package:go_router/go_router.dart';
 
@@ -16,6 +17,7 @@ part 'route.g.dart';
     TypedGoRoute<CustomUsageRoute>(path: '/custom-usage'),
     TypedGoRoute<NestedScrollRoute>(path: '/nested-scroll'),
     TypedGoRoute<AdaptiveSizingRoute>(path: '/adaptive-sizing'),
+    TypedGoRoute<TransitionBuilderRoute>(path: '/transition-builder'),
     TypedGoRoute<ParallaxEffectRoute>(path: '/parallax-effect'),
   ],
 )